Remove submodule checkout in build.rs
Crates on crates.io are not pulled as git repositories. This causes
a) the path libui/.git to never exist, and, subsequently
b) git erroring out with:

```
Caused by:
  process didn't exit successfully: `libui-rs-static/target/debug/build/libui-sys-086734ec98b21422/build-script-build` (exit code: 101)
--- stderr
fatal: not a git repository (or any parent up to mount point /)
Stopping at filesystem boundary (GIT_DISCOVERY_ACROSS_FILESYSTEM not set).
```

This error is usually silently swallowed, but crops up when the build fails
for other reasons (e.g. build without meson installed).

For dependent crates in cargo builds this change is backwards compatible, since
cargo/crates.io checks out submodules by default. For direct builds from the
git source the submodule checkout should be at the descretion of the user. This
also removes the git dependency in the build.